+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 0.3.6 (2020-02-24)
+
+Features:
+  - Test Run view: Add stability indicator -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/60)
+  - Provide public API -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/56)
+
 ## 0.3.5 (2020-02-15)
 
 Features: